add routes

+++ internal/repository/team.go
@@ -7,9 +7,9 @@ import (
 
 type TeamRepository interface {
 	Create(team *models.Team) error
-	GetByID(id int) (*models.Team, error)
+	GetByID(id uint64) (*models.Team, error)
 	GetOpenTeams() ([]*models.Team, error)
-	Update(id int, team *models.Team) error
+	Update(id uint64, team *models.Team) error
 }
 
 type TeamRepositoryImpl struct {
@@ -24,7 +24,7 @@ func (t *TeamRepositoryImpl) Create(team *models.Team) error {
 	return t.db.Create(team).Error
 }
 
-func (t *TeamRepositoryImpl) GetByID(id int) (*models.Team, error) {
+func (t *TeamRepositoryImpl) GetByID(id uint64) (*models.Team, error) {
 	var team models.Team
 	err := t.db.Where("id = ?", id).First(&team).Error
 	if err == gorm.ErrRecordNotFound {
@@ -44,6 +44,6 @@ func (t *TeamRepositoryImpl) GetOpenTeams() ([]*models.Team, error) {
 	}
 }
 
-func (t *TeamRepositoryImpl) Update(id int, team *models.Team) error {
+func (t *TeamRepositoryImpl) Update(id uint64, team *models.Team) error {
 	return t.db.Model(&models.Team{}).Where("id = ?", id).Updates(team).Error
 }
